Split MemberListItem and extract it's shared sub components
