+++ packages/ui/src/memberships/components/MemberInfo.tsx
@@ -29,10 +29,21 @@ interface MemberInfoContainerProps {
   maxRoles?: number
 }
 
-export type MemberInfoProps = MemberInfoContainerProps & MemberInfoWrapProps
+export type MemberInfoProps = MemberInfoContainerProps & MemberInfoWrapProps & { showGroup?: boolean }
 
 export const MemberInfo = React.memo(
-  ({ member, onClick, isOnDark, showId, memberSize, size, className, maxRoles, isLeader }: MemberInfoProps) => {
+  ({
+    member,
+    onClick,
+    isOnDark,
+    showId,
+    showGroup = true,
+    memberSize,
+    size,
+    className,
+    maxRoles,
+    isLeader,
+  }: MemberInfoProps) => {
     return (
       <MemberInfoWrap isOnDark={isOnDark} memberSize={memberSize} className={className}>
         <MemberPhoto>
@@ -48,7 +59,7 @@ export const MemberInfo = React.memo(
             {(member as any)?.isFounder && <MemberStatusHelp icon={<FounderMemberIcon />} helperText="Lorem fishy" />}
           </MemberIcons>
         </MemberHead>
-        {member.type !== 'Members' && !showId && <MemberRoles member={member} size={size} max={maxRoles} />}
+        {showGroup && !showId && <MemberRoles member={member} size={size} max={maxRoles} />}
         {showId && <MemberId>Worker ID: {member.id}</MemberId>}
       </MemberInfoWrap>
     )
