+++ packages/ui/src/memberships/components/MemberListItem/index.ts
@@ -1,0 +1,2 @@
+export * from './Member'
+export * from './MyMember'
